Refaktor tekster
