--- src/components/melding/Melding.less
@@ -1,3 +1,0 @@
-.message {
-  margin: 1rem 0;
-}
