+++ src/pages/digital-samhandling-offentlig-privat/DsopHistorikk.tsx
@@ -10,6 +10,7 @@ import { Link, useLocation } from "react-router-dom";
 import PageContainer from "components/pagecontainer/PageContainer";
 import DSOPIkon from "assets/img/DSOP.svg";
 import WithDSOP from "./DsopFetch";
+import Lenke from "nav-frontend-lenker";
 
 interface Props {
   dsopInfo: DsopInfo;
@@ -53,7 +54,7 @@ const Tabell = (props: Props) => {
     if (!initState[year]) {
       initState[year] = {
         innslag: [innslag],
-        ekspandert: !i
+        ekspandert: !i,
       };
     } else {
       initState[year].innslag.push(innslag);
@@ -68,7 +69,16 @@ const Tabell = (props: Props) => {
         <>
           <div className={"historikk__info"}>
             <AlertStripeInfo>
-              <FormattedMessage id={"lenker.dsop.info"} />
+              <FormattedMessage
+                id={"lenker.dsop.info"}
+                values={{
+                  a: (text: String) => (
+                    <Lenke href="https://altinn.no" target="blank">
+                      {text}
+                    </Lenke>
+                  ),
+                }}
+              />
             </AlertStripeInfo>
           </div>
           <div className="historikk__flex-rad historikk__head">
@@ -85,15 +95,15 @@ const Tabell = (props: Props) => {
           </div>
           {Object.keys(data)
             .reverse()
-            .map(year => {
+            .map((year) => {
               const value = data[year];
               const onClick = () =>
                 setData({
                   ...data,
                   [year]: {
                     ...data[year],
-                    ekspandert: !data[year].ekspandert
-                  }
+                    ekspandert: !data[year].ekspandert,
+                  },
                 });
 
               return (
