+++ src/pages/forside/sections/7-mer-informasjon/MerInformasjonData.ts
@@ -2,21 +2,21 @@ export default [
   {
     id: 1,
     tittel: "alternativer.feilopplysninger.tittel",
-    melding: "alternativer.feilopplysninger.melding"
+    melding: "alternativer.feilopplysninger.melding",
   },
   {
     id: 2,
     tittel: "alternativer.savneropplysninger.tittel",
-    melding: "alternativer.savneropplysninger.melding"
+    melding: "alternativer.savneropplysninger.melding",
   },
   {
     id: 3,
     tittel: "alternativer.utlevere.tittel",
-    melding: "alternativer.utlevere.melding"
+    melding: "alternativer.utlevere.melding",
   },
   {
     id: 4,
     tittel: "alternativer.klage.tittel",
-    melding: "alternativer.klage.melding"
-  }
+    melding: "alternativer.klage.melding",
+  },
 ];
