+++ src/pages/skattetrekksmelding/SkattHistorikk.tsx
@@ -10,6 +10,7 @@ import { InstInfo } from "types/inst";
 import PageContainer from "components/pagecontainer/PageContainer";
 import SkattIkon from "assets/img/Skattekort.svg";
 import WithSkatt from "./SkattFetch";
+import Lenke from "nav-frontend-lenker";
 
 /*
   Hent data
@@ -49,7 +50,7 @@ const Tabell = (props: { instInfo: InstInfo }) => {
     if (!initState[year]) {
       initState[year] = {
         instInnslag: [instInnslag],
-        ekspandert: !i ? true : false
+        ekspandert: !i ? true : false,
       };
     } else {
       initState[year].instInnslag.push(instInnslag);
@@ -64,7 +65,16 @@ const Tabell = (props: { instInfo: InstInfo }) => {
         <>
           <div className={"historikk__info"}>
             <AlertStripeInfo>
-              <FormattedMessage id={"lenker.dsop.info"} />
+              <FormattedMessage
+                id={"lenker.dsop.info"}
+                values={{
+                  a: (text: String) => (
+                    <Lenke href="https://altinn.no" target="blank">
+                      {text}
+                    </Lenke>
+                  ),
+                }}
+              />
             </AlertStripeInfo>
           </div>
           <div className="historikk__flex-rad historikk__head">
@@ -81,15 +91,15 @@ const Tabell = (props: { instInfo: InstInfo }) => {
           </div>
           {Object.keys(data)
             .reverse()
-            .map(year => {
+            .map((year) => {
               const value = data[year];
               const onClick = () =>
                 setData({
                   ...data,
                   [year]: {
                     ...data[year],
-                    ekspandert: !data[year].ekspandert
-                  }
+                    ekspandert: !data[year].ekspandert,
+                  },
                 });
 
               return (
