+++ src/text/nb.ts
@@ -22,7 +22,7 @@ export default {
   "brodsmulesti.arbeidsforhold": `Arbeidsforhold`,
   "header.hello": `Hei!`,
   "header.hello.name": `Hei {name}!`,
-  "header.description": `<a href='https://www.nav.no/personvern' target='blank' rel='noopener noreferrer' class='lenke'>Les om personvern i Arbeids- og velferdsetaten her.</a>`,
+  "header.description": `<a>Les om personvern i Arbeids- og velferdsetaten her.</a>`,
   "header.obs": `Her kan du se opplysninger som NAV har registrert om deg. Du har rett til innsyn i personopplysninger NAV har registrert om deg, men det er ikke sikkert du finner alle personopplysninger her. Det kan skyldes både tekniske og juridiske begrensninger.`,
   "personalia.tittel": `Personalia`,
   "personalia.beskrivelse": `Vi må vite hvem du er og hvordan vi kan kontakte deg. Hvis du skal ha tjenester eller søker om økonomisk støtte, trenger vi opplysninger for å kunne vurdere hva du har krav på. For å kunne utbetale penger trenger vi kontonummeret ditt og en adresse.`,
@@ -54,30 +54,29 @@ export default {
   "personalia.iban": `IBAN`,
   "personalia.kontonrelleriban": `Kontonummer / IBAN`,
   "personalia.kontonr.oveskrift": `Kontonummer vi utbetaler til`,
-  "personalia.kontonr.ingenData": `Vi har ikke registrert kontonummer på deg. <br>Du kan registrere kontonummer ved å trykke på «Legg til»`,
+  "personalia.kontonr.ingenData": `Vi har ikke registrert kontonummer på deg. <br></br>Du kan registrere kontonummer ved å trykke på «Legg til»`,
   "personalia.bank": `Bank`,
   "personalia.valuta": `Valuta`,
   "personalia.bickode": `BIC / SWIFT-kode`,
   "personalia.bankkode": `Bankkode`,
   "personalia.tlfnr": `Telefon`,
   "personalia.tlfnr.oveskrift": `Telefonnummer registrert hos NAV`,
-  "personalia.tlfnr.ingenData": `Vi har ikke registrert telefonnummer på deg i våre registre. <br>Har du et telefonummer vi kan kontakte deg på?`,
+  "personalia.tlfnr.ingenData": `Vi har ikke registrert telefonnummer på deg i våre registre. <br></br>Har du et telefonummer vi kan kontakte deg på?`,
   "personalia.tlfnr.telefon": `Telefon {x}`,
   "personalia.tlfnr.slett.alert": `Er du sikker på at du vil fjerne dette nummeret?`,
   "personalia.dkif.overskrift": `Opplysninger fra Kontakt - og reservasjonsregisteret`,
   "personalia.dkif.beskrivelse": `Kontakt- og reservasjonsregisteret er et felles nasjonalt register. NAV må ha dine kontaktopplysninger i dette registeret for å kunne sende deg meldinger og dokumenter elektronisk.`,
   "personalia.dkif.kanVarsles": `Du har <b>ikke</b> reservert deg mot elektronisk kommunikasjon.`,
   "personalia.dkif.kanIkkeVarsles": `Du har reservert deg mot elektronisk kommunikasjon.`,
-  "personalia.dkif.ingenData": `Ingen opplysninger registrert`,
+  "personalia.dkif.ingenData": `<b>Ingen opplysninger registrert<b>`,
   "adresse.tittel": `Adresser`,
   "adresse.beskrivelse": `Her finner du bostedsadresse og eventuelt din midlertidige adresse.`,
   "adresse.bostedsadresse": `Bostedsadresse`,
   "adresse.bostedsadresse.apneTekst": `Vis detaljer`,
   "adresse.bostedsadresse.lukkTekst": `Skjul detaljer`,
   "adresse.midlertidigadresse": `Midlertidig adresse`,
-  "adresse.midlertidigadresse.leggtil.beskrivelse": `Du har ikke registrert en midlertidig adresse. <br>Hvis du ønsker å få post fra NAV til en midlertidig adresse, kan du registrere adressen.`,
+  "adresse.midlertidigadresse.leggtil.beskrivelse": `Du har ikke registrert en midlertidig adresse. <br></br>Hvis du ønsker å få post fra NAV til en midlertidig adresse, kan du registrere adressen.`,
   "adresse.utenlandskadresse": `Midlertidig adresse i utlandet`,
-  "adresse.utenlandskadresse.leggtil.beskrivelse": `Du har ikke registrert en midlertidig adresse. <br>Hvis du ønsker å få post fra NAV til en midlertidig adresse, kan du registrere adressen. `,
   "adresse.postadresse": `Postadresse`,
   "adresse.adressetillegg": `Adressetillegg`,
   "adresse.gate": `Gate`,
@@ -90,7 +89,7 @@ export default {
   "adresse.overskrift": `Adresser som er registert i Folkeregisteret`,
   "adresse.ikke.registrert": `Du har ikke registrert bostedsadresse i Folkeregisteret`,
   "adresse.midlertidig.alert": `Frem til og med {dato} sender vi deg post til din midlertidige adresse.`,
-  "adresse.slett.alert": `Vil du fjerne den midlertidige adressen og bruke folkeregistrert </br> adresse til å motta post fra NAV?`,
+  "adresse.slett.alert": `Vil du fjerne den midlertidige adressen og bruke folkeregistrert <br></br> adresse til å motta post fra NAV?`,
   "adresse.hjelpetekster.adressetyper": `<p>Postadresse inneholder gatenavn, gatenummer og postnummer.</p><p>Postboksadresse viser til en postboks på et postkontor eller post i butikk.</p><p>Stedsadresse viser til et sted, for eksempel en gård, og inneholder ikke gatenavn</p>`,
   "adresse.hjelpetekster.gyldigtil": `<p>Du velger selv hvor lenge adressen skal være gyldig, maksimalt 1 år. Etter 1 år må du endre eller forlenge adressen.</p>`,
   "adresse.hjelpetekster.co": `<p>Bor du midlertidig på en annen persons adresse og ønsker posten dit? Da må posten være merket med den andre personens navn. Det er ekstra viktig hvis navnet ditt ikke står på postkassen.</p>`,
@@ -98,7 +97,7 @@ export default {
   "adresse.hjelpetekster.utenlandsk.adresse": `<p>Vær nøye med å fylle ut riktig adresse slik at posten kommer fram. Du kan dele adressen over flere linjer, inntil 30 tegn per linje.</p>`,
   "arbeidsforhold.tittel": `Arbeidsforhold`,
   "arbeidsforhold.beskrivelse": `<p>NAV trenger informasjon om arbeidsforholdene dine for å kunne vurdere behovet ditt for oppfølging og tjenester fra NAV, og til å beregne hvor mye du kan få utbetalt i for eksempel sykepenger, dagpenger og foreldrepenger. Andre offentlige etater har også tilgang til opplysninger fra Arbeidsgiver og arbeidstakerregisteret.</p><p>Det er Arbeids- og velferdsetaten som eier og forvalter Arbeidsgiver og arbeidstakerregisteret (Aa-registeret).<br/><a href="https://www.nav.no/no/Bedrift/Tjenester+og+skjemaer/Aa-registeret+og+a-meldingen">Her finner du mer informasjon om Arbeidsgiver- og arbeidstakerregisteret</a>.</p>`,
-  "arbeidsforhold.kilde": `Kilde: Arbeidsgiver- og arbeidstakerregisteret</br><span style="text-transform: none;">Dataene er rapportert av arbeidsgiver</span>`,
+  "arbeidsforhold.kilde": `Kilde: Arbeidsgiver- og arbeidstakerregisteret<br></br><span>Dataene er rapportert av arbeidsgiver</span>`,
   "arbeidsforhold.disclaimer": `Oversikten viser alle arbeidsforhold rapportert etter 01.01.2015.<br/>Hvis det er feil i et arbeidsforhold, kan arbeidsgiveren endre gjennom a-meldingen.`,
   "arbeidsforhold.submitted.by": `Dataene er rapportert av arbeidsgiver`,
   "utbetalinger.tittel": "Kontonummer for utbetaling",
@@ -168,10 +167,9 @@ export default {
   "lenker.cv.beskrivelse": `Se CV-en og jobbprofilen din, og dine lagrede søk på stillinger.`,
   "lenker.dsop.tittel": `Utlevering av personopplysninger`,
   "lenker.dsop.beskrivelse": `Se hvilke opplysninger NAV kan gi til forsikringsselskap, fordi du har gitt samtykke.`,
-  "lenker.dsop.info": `Dette er en logg som viser hvilke opplysninger vi har utlevert til forsikringsselskap basert på samtykke du har gitt på Altinn. <a href="https://altinn.no">Se samtykket i Altinn.</a>`,
+  "lenker.dsop.info": `Dette er en logg som viser hvilke opplysninger vi har utlevert til forsikringsselskap basert på samtykke du har gitt på Altinn. <a>Se samtykket i Altinn.</a>`,
   "lenker.inst.tittel": `Institusjonsopphold`,
   "lenker.inst.beskrivelse": `Se dine institusjonsopphold`,
-  "lenker.inst.info": `Dette er en logg som viser hvilke opplysninger vi har utlevert til forsikringsselskap basert på samtykke du har gitt på Altinn. <a href="https://altinn.no">Se samtykket i Altinn.</a>`,
   "lenker.skattetrekksmelding.tittel": `Skattetrekksmeldinger`,
   "lenker.skattetrekksmelding.beskrivelse": `Se dine skattetrekksmeldinger`,
   "lenker.medl.tittel": `Unntak fra medlemskap i folketrygden`,
@@ -211,7 +209,7 @@ export default {
   "felter.kontonummer.label": "Kontonummer",
   "felter.kontonummervalg.norsk": "Norsk kontonummer",
   "felter.kontonummervalg.utenlandsk": "Utenlandsk kontonummer",
-  "felter.utenlandskkonto.info": `Vær nøye med utfylling av bankinformasjon for utenlandsk konto slik at betalingen kommer frem til riktig sted. NAV er ikke ansvarlig for at opplysningene er riktige. Gebyrer kan påløpe. Les mer om krav til <a href="https://www.nav.no/no/NAV+og+samfunn/Kontakt+NAV/Utbetalinger/Utbetalinger/utbetaling-av-ytelser-fra-nav-til-utlandet"> utenlandsk kontonummer </a>.`,
+  "felter.utenlandskkonto.info": `Vær nøye med utfylling av bankinformasjon for utenlandsk konto slik at betalingen kommer frem til riktig sted. NAV er ikke ansvarlig for at opplysningene er riktige. Gebyrer kan påløpe. Les mer om krav til <a> utenlandsk kontonummer </a>.`,
   "felter.kontonummer.kontonummer.label": "Kontonummer / IBAN",
   "felter.bic.label": "BIC / Swift-kode",
   "felter.bankensland.label": "Bankens land",
@@ -222,7 +220,7 @@ export default {
   "felter.bankidentifier.bankkode": "Bankkode",
   "felter.bankidentifier.harikke.bic": `Har ikke BIC / Swift-kode`,
   "felter.bankidentifier.harikke.bic.advarsel": `Dersom du ikke oppgir BIC / Swift-kode kan utbetalingen bli forsinket eller avvist.`,
-  "felter.landetbrukerbankkode.advarsel": `For <span class="capitalize">{land}</span> kan du oppgi både BIC/Swift-kode og bankkode. Du kan oppgi begge, men det er tilstrekkelig med en av kodene.`,
+  "felter.landetbrukerbankkode.advarsel": `For <span>{land}</span> kan du oppgi både BIC/Swift-kode og bankkode. Du kan oppgi begge, men det er tilstrekkelig med en av kodene.`,
   "validation.only.letters": "Skriv bare bokstaver",
   "validation.only.digits": "Skriv bare tall",
   "validation.only.digits.space": "Skriv bare tall og mellomrom",
@@ -277,5 +275,5 @@ export default {
   "alternativer.utlevere.tittel": merInfo.utlevere.tittel,
   "alternativer.utlevere.melding": merInfo.utlevere.melding,
   "alternativer.klage.tittel": merInfo.klage.tittel,
-  "alternativer.klage.melding": merInfo.klage.melding
+  "alternativer.klage.melding": merInfo.klage.melding,
 };
