+++ src/text/nb/informasjon.ts
@@ -6,28 +6,28 @@ export default {
       `<p>Hvis personopplysninger om deg er feil, kan du be om at de blir rettet. Hvis personopplysninger er ufullstendige, kan du be om at de blir supplert. ` +
       `I særlige tilfeller kan du ha rett til å få slettet personopplysninger om deg, men NAV har som regel lovbestemt plikt til å lagre opplysningene.</p>` +
       `<p>Hvis personopplysningene er hentet fra andre kilder enn NAV, kan du henvende deg til kilden som er oppgitt.</p>` +
-      `<p>Hvis kilden er NAV, kan du bruke <a class="lenke" href="https://www.nav.no/beskjedtilnav" target="_blank" rel="noopener noreferrer">Send beskjed til NAV</a> eller henvende deg til NAV Kontaktsenter, tlf. <a class="lenke" href="tel:004755553333">55 55 33 33</a>.</p>`
+      `<p>Hvis kilden er NAV, kan du bruke <beskjedLenke>Send beskjed til NAV</beskjedLenke> eller henvende deg til NAV Kontaktsenter, tlf. <telefonLenke>55 55 33 33</telefonLenke>.</p>`,
   },
   savneropplysninger: {
     tittel: `Savner du personopplysninger?`,
     melding:
       `<p>Denne siden er under utvikling. Derfor er det ikke sikkert du finner alle personopplysninger NAV har om deg her. ` +
       `Det kan være både tekniske og juridiske begrensninger som gjør at vi ikke kan vise personopplysninger. </p>` +
-      `<p>Savner du personopplysninger eller lurer du på om NAV har flere personopplysninger om deg, kan du bruke <a class="lenke" href="https://www.nav.no/beskjedtilnav" target="_blank" rel="noopener noreferrer">Send beskjed til NAV</a> eller henvende deg til NAV Kontaktsenter, tlf. <a class="lenke" href="tel:004755553333">55 55 33 33</a>.</p>`
+      `<p>Savner du personopplysninger eller lurer du på om NAV har flere personopplysninger om deg, kan du bruke <beskjedLenke>Send beskjed til NAV</beskjedLenke> eller henvende deg til NAV Kontaktsenter, tlf. <telefonLenke>55 55 33 33</telefonLenke>.</p>`,
   },
   utlevere: {
     tittel: `Utlevere eller innhente opplysninger til/fra andre land`,
     melding:
       `<p>Hvis du har bodd eller arbeidet i andre land enn Norge, og har søkt om ytelser eller tjenester i Norge eller i et annet land,  må NAV noen ganger utveksle opplysninger om deg med et annet lands trygdemyndigheter. ` +
-      `NAV vil bare levere ut opplysninger til land som er omfattet av EØS-avtalen eller land som Norge har inngått trygdeavtaler med. ` +
-      `<p><a class="lenke" href="https://lovdata.no/nav/andre-rettskilder/Trygdeavtaler" target="blank" rel="noopener noreferrer">Se oversikt over hvilke land Norge har inngått trygdeavtale med</a>.</p>`
+      `NAV vil bare levere ut opplysninger til land som er omfattet av EØS-avtalen eller land som Norge har inngått trygdeavtaler med. </p>` +
+      `<p><trygdeavtalerLenke>Se oversikt over hvilke land Norge har inngått trygdeavtale med</trygdeavtalerLenke>.</p>`,
   },
   klage: {
     tittel: `Har du spørsmål eller ønsker du å klage?`,
     melding:
-      `<p>Hvis du har spørsmål om vår behandling av dine personopplysninger, kan du bruke <a class="lenke" href="https://www.nav.no/beskjedtilnav" target="_blank" rel="noopener noreferrer">Send beskjed til NAV</a> eller henvende deg til NAV Kontaktsenter, tlf. <a class="lenke" href="tel:004755553333">55 55 33 33</a>.</p>` +
+      `<p>Hvis du har spørsmål om vår behandling av dine personopplysninger, kan du bruke <beskjedLenke>Send beskjed til NAV</beskjedLenke> eller henvende deg til NAV Kontaktsenter, tlf. <telefonLenke>55 55 33 33</telefonLenke>.</p>` +
       `<p>Du har rett til å klage til Datatilsynet hvis du ikke er fornøyd med hvordan vi behandler dine personopplysninger, eller hvis du mener behandlingen er i strid med personvernreglene. ` +
-      `Informasjon om hvordan du går frem finner du på nettsidene til <a class="lenke" href="https://www.datatilsynet.no/om-datatilsynet/kontakt-oss/hvordan-kan-jeg-klage-til-datatilsynet" target="_blank" rel="noopener noreferrer">Datatilsynet</a>. </p>` +
-      `<p>Vi oppfordrer deg til å kontakte NAV før du klager til Datatilsynet, slik at vi kan vurdere spørsmålene dine og informere deg om hvordan vi behandler personopplysninger.</p>`
-  }
+      `Informasjon om hvordan du går frem finner du på nettsidene til <datatilsynetLenke>Datatilsynet</datatilsynetLenke>. </p>` +
+      `<p>Vi oppfordrer deg til å kontakte NAV før du klager til Datatilsynet, slik at vi kan vurdere spørsmålene dine og informere deg om hvordan vi behandler personopplysninger.</p>`,
+  },
 };
